refactor: move code to hook
